vector_db_project/requirements.txt
transformers
sentence-transformers
faiss-cpu
PyMuPDF
numpy
